Made an Interactable "Lever"
If you go to the capsule (named lever) at the right of the play field, and press e, it displays "Do something..." in the console. If there was going to be something done, It would be done. This is done using the InteractionObject script, which can be attached to any GameObject with a  Circle Collider 2D, and that GameObject will do something when E is pressed.